Many DB Classes added. Main page in better condition, other pages added

+++ src/java/az/qutan/model/BloodGroup.java
@@ -1,12 +1,4 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
 package az.qutan.model;
-
-import java.math.BigDecimal;
 
 /**
  *
@@ -25,7 +17,7 @@ public class BloodGroup extends Base{
         this.rhesus = "";
     }
 
-    public BloodGroup(BigDecimal id, String name, String rhesus, int status) {
+    public BloodGroup(int id, String name, String rhesus, int status) {
         super(id, status);
         this.name = name;
         this.rhesus = rhesus;
